requirements.txt
Flask==3.0.0
gunicorn==20.0.4
line-bot-sdk==3.7.0
psycopg2==2.9.9
APScheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2